Make User and Group Serializable
Fixes #129

The classes `User` and `Group` implement the interface `Serializable` and have a
`serialVersionUID` field. This is also true for the classes that are
used for fields of `User` and `Group`.

A test for each of the classes tests that no exception is thrown when
serializing and deserializing the object, and that the deserialized is
equal to the original.