Restore compatibility with old Jackson versions
Jackson 2.5 introduced new methods to `TypeFactory` that the connector
is using to deserialize a SCIM search result. Not everybody has updated
to Jackson >= 2.5, so incompatibilities have arisen. This commit
restores the usage of the old, but deprecated method to construct a
proper `JavaType` instance. It will call the new method anyway.