Remove distinction between auth- and resource-server
The core servers are merged now, so the connector has to catch up with
that. Instead of 3 separate methods for setting endpoints, there will be
only 1 left, namely `OsiamConnector#setEndpoint`, to set the endpoint of
OSIAM. Also removed the ability to configure the maximum connections per
route, because there will be only 1 route from now on.