Change version to 'latest-SNAPSHOT'
This has to be done for the new release process to work. Otherwise we
would have to increase the version somehow after release by Circle CI or
worse: manually. There is no need to fetch a snapshot after a release,
because the release contains the exact same code as the last snapshot.
So having a universally working snapshot version, like
'latest-SNAPSHOT', should do no harm, but eases the burden of
maintaining the right version between releases.