Update deprecations from scim-schema
With the merge of the scim-schema, deprecations have also been merged,
but were not aligned to the version of the connector. The alignment is
as follows: scim-schema 1.6 is the same as connector 1.8, so all "removal
versions" have to be increased by 2. Update deprecations without a
removal version on a best effort basis.